Added total packets count on quit

+++ ddos.h
@@ -22,7 +22,7 @@
 #define MODE_TCP 1
 #define MODE_UDP 2
 
-extern int pc;
+extern int64_t pc;
 extern int tcount;
 extern char* __host;
 extern int __port;
